fix getintensity method return value #BUILD

+++ Readers/E57.cs
@@ -154,7 +154,7 @@ namespace PointCloudConverter.Readers
             return default;
         }
 
-        public byte GetIntensity()
+        public ushort GetIntensity()
         {
             var i = currentPointIndex - 1;
             if (currentChunk?.Intensities != null && i >= 0 && i < currentChunk.Intensities.Length)
